Add custom fuzzy sorting and typing to data table
Introduced fuzzy sorting functionality by defining custom typing and implementing a `fuzzySort` method in the data table. Centralized type definitions into a separate file for better type management and removed redundant inline module declarations.
