+++ packages/core-ui/src/@types/data-table.d.ts
@@ -1,0 +1,15 @@
+import { RankingInfo } from '@tanstack/match-sorter-utils';
+
+declare module '@tanstack/react-table' {
+  interface FilterFns {
+    fuzzy: FilterFn<unknown>;
+  }
+
+  interface SortingFns {
+    fuzzy: SortingFn<unknown>;
+  }
+
+  interface FilterMeta {
+    itemRank: RankingInfo;
+  }
+}
