+++ packages/core-ui/src/components/ui/data-table/data-table.tsx
@@ -80,6 +80,9 @@ const DataTable = Object.assign(
       filterFns: {
         fuzzy: fuzzyFilter,
       },
+      sortingFns: {
+        fuzzy: fuzzySort,
+      },
     });
 
     return (
